lambdas/mix-ai-email/requirements.txt
boto3
google-generativeai
requests


